get list of reasons from database
I know you said I should create a new servlet, but I couldn't figure out to do know, so I used req.getPathInfo to get the path after the base url. Then a switch to choose the right method in the DAO. I know you showed a way to use a substring to get the path after the base url, but I googled a better way.

I added a state variable to hold the reasons, then added map functions to output them on the page instead of hardcoding them